search cpf and return indexOf feature
