+++ src/user/UserRepo.java
@@ -178,6 +178,18 @@ public class UserRepo {
         }
         return retorno;
     }
+    
+    public int verificaExistencia2(User user) {
+        int retorno = -1;
+        for (int i = 0; i < this.lista.size(); i++) {
+            if (user.getCpf().trim().equals(this.lista.get(i).getCpf().trim())) {
+                retorno = i;
+                break;
+            }
+        }
+        JOptionPane.showMessageDialog(null, retorno);
+        return retorno;
+    }
 
     
     public class Vetor {
